refactored metadata to split files by node type added metadata and component registry objects that manage all available class types.
