+++ Freeform.Python/DCCUtilities/Maya/maya_utils/node_utils.py
@@ -110,7 +110,7 @@ def zero_node(obj, connection_filter_list):
     locked_attrs = unlock_transforms(obj)
     
     tr_zero_value_list = [0,0,0,0,0,0]
-    control_property = metadata.meta_properties.get_property(obj, metadata.meta_properties.ControlProperty)
+    control_property = metadata.meta_property_utils.get_property(obj, metadata.meta_properties.ControlProperty)
     
     if control_property:
         zero_translate = control_property.get('zero_translate', 'double3')
@@ -230,7 +230,7 @@ def flip_if_world(control_obj, axis, flip_method):
         axis (str): String name for the axis to flip on, 'x', 'y', or 'z'
         flip_method (method): Should be either flip_attribute_keys or flip_transforms
     '''
-    control_property = metadata.meta_properties.get_property(control_obj, metadata.meta_properties.ControlProperty)
+    control_property = metadata.meta_property_utils.get_property(control_obj, metadata.meta_properties.ControlProperty)
     is_world = control_property.get('world_space', 'bool') if control_property else False
     if is_world:
         flip_method(control_obj, v1_shared.shared_utils.get_mirror_attributes(axis))
