+++ Freeform.Python/DCCUtilities/Maya/maya_utils/usertools/aim_constraint_dialogue.py
@@ -112,7 +112,7 @@ class Aim_Constraint_Dialogue(object):
             aim_space = pm.spaceLocator(n=target.name() + "_temp_aimspace_target")
             aim_space.localScale.set(10,10,10)
 
-            character_network = metadata.network_core.MetaNode.get_first_network_entry(target, metadata.network_core.CharacterCore)
+            character_network = metadata.meta_network_utils.get_first_network_entry(target, metadata.network_core.CharacterCore)
             if character_network:
                 scalar = character_network.get('scalar', 'float')
                 scalar = 1 if not scalar else scalar
