+++ Freeform.Python/FreeformRigging/Maya/freeform_utils/character_utils.py
@@ -53,7 +53,7 @@ def characterize_skeleton(jnt, name = None, update_ui = True, freeze_skeleton = 
 
     rigging.skeleton.clean_skeleton(jnt)
 
-    character_network = metadata.network_core.MetaNode.get_first_network_entry(jnt, metadata.network_core.CharacterCore)
+    character_network = metadata.meta_network_utils.get_first_network_entry(jnt, metadata.network_core.CharacterCore)
     if character_network:
         v1_shared.usertools.message_dialogue.open_dialogue("This Skeleton is already characterized -- {0}".format(character_network.character_name.get()), title="Already Characterized")
         return
@@ -168,22 +168,22 @@ def characterize_with_zeroing(jnt = None):
     scene_tools.scene_manager.SceneManager().run_by_string('UpdateRiggerInPlace')
 
 
-def remove_existing_rigging(component_type, joint_chain, force_remove = False):
+def remove_existing_rigging(has_attachment, joint_chain, force_remove = False):
     local_bake_queue = maya_utils.baking.BakeQueue("Remove Existing Queue")
     character_category = v1_core.global_settings.GlobalSettings().get_category(v1_core.global_settings.CharacterSettings)
     if character_category.remove_existing or force_remove:
         for region_joint in joint_chain[1:-1]:
             rigging.rig_base.Component_Base.remove_rigging(region_joint, local_queue=local_bake_queue)
-        if component_type._hasattachment != 'root':
+        if has_attachment != 'root':
             removed_node_list = rigging.rig_base.Component_Base.remove_rigging(joint_chain[-1], exclude = 'end', local_queue=local_bake_queue)
-        if component_type._hasattachment != 'end':
+        if has_attachment != 'end':
             removed_node_list = rigging.rig_base.Component_Base.remove_rigging(joint_chain[0], exclude = 'root', local_queue=local_bake_queue)
 
         local_bake_queue.run_queue()
 
 
 def transfer_ue_anim_to_character(character_node, ue4_animation_path):
-    character_network = metadata.network_core.MetaNode.create_from_node(character_node)
+    character_network = metadata.meta_network_utils.create_from_node(character_node)
     character_skeleton = character_network.get_downstream(metadata.network_core.JointsCore).get_connections()
 
     ue_settings_path = None
@@ -210,8 +210,8 @@ def update_rig_file():
     Update a rig file from an FBX that the user is prompted to choose.  Keeps current rig settings and inserts the updated
     character skeleton/model from the chosen FBX file.
     '''
-    current_character_node = metadata.network_core.MetaNode.get_all_network_nodes(metadata.network_core.CharacterCore)[0]
-    current_character_network = metadata.network_core.MetaNode.create_from_node(current_character_node)
+    current_character_node = metadata.meta_network_utils.get_all_network_nodes(metadata.network_core.CharacterCore)[0]
+    current_character_network = metadata.meta_network_utils.create_from_node(current_character_node)
     character_namespace = current_character_network.group.namespace()
     character_name = current_character_network.node.character_name.get()
     character_root_path = current_character_network.node.root_path.get()
@@ -328,7 +328,7 @@ def mirror_center_pose(region_network, axis):
 
 
 def get_mirror_network_from_control(control_obj, mirror_dict):
-    component_network = metadata.network_core.MetaNode.get_first_network_entry(control_obj, metadata.network_core.ComponentCore)
+    component_network = metadata.meta_network_utils.get_first_network_entry(control_obj, metadata.network_core.ComponentCore)
     mirror_network = get_matching_component(component_network, mirror_dict)
 
     return component_network, mirror_network
@@ -349,7 +349,7 @@ def mirror_pose_matching_regions(source_network, mirror_network, axis, single_di
     source_control_list, mirror_control_list = get_matching_region_controls(source_network, mirror_network)
 
     for source_control, mirror_control in zip(source_control_list, mirror_control_list):
-        control_property = metadata.meta_properties.get_property(source_control, metadata.meta_properties.ControlProperty)
+        control_property = metadata.meta_property_utils.get_property(source_control, metadata.meta_properties.ControlProperty)
         is_world = control_property.get('world_space', 'bool') if control_property else False
         if is_world:
             maya_utils.node_utils.world_space_mirror(source_control, mirror_control, axis, single_direction)
@@ -407,11 +407,11 @@ def get_matching_region_controls(source_network, mirror_network):
 
 def get_matching_control(control_obj, mirror_dict):
     found_match = False
-    overdriven_network_entry = metadata.network_core.MetaNode.get_first_network_entry(control_obj, metadata.network_core.AddonCore)
+    overdriven_network_entry = metadata.meta_network_utils.get_first_network_entry(control_obj, metadata.network_core.AddonCore)
 
-    control_property = metadata.meta_properties.get_property(control_obj, metadata.meta_properties.ControlProperty)
+    control_property = metadata.meta_property_utils.get_property(control_obj, metadata.meta_properties.ControlProperty)
     control_index = control_property.get('ordered_index')
-    component_network = metadata.network_core.MetaNode.get_first_network_entry(control_obj, metadata.network_core.ComponentCore)
+    component_network = metadata.meta_network_utils.get_first_network_entry(control_obj, metadata.network_core.ComponentCore)
 
     mirror_network = get_matching_component(component_network, mirror_dict)
 
@@ -421,7 +421,7 @@ def get_matching_control(control_obj, mirror_dict):
         mirror_control_list = mirror_control_network.get_connections()
 
         for check_control in mirror_control_list:
-            mirror_property = metadata.meta_properties.get_property(check_control, metadata.meta_properties.ControlProperty)
+            mirror_property = metadata.meta_property_utils.get_property(check_control, metadata.meta_properties.ControlProperty)
             if mirror_property.get('ordered_index') == control_index:
                 mirror_control = check_control
                 found_match = True
@@ -486,7 +486,7 @@ def get_addon_from_control(control_obj):
     '''
     Get the Addon control object from the give rig control object if it's overdriven
     '''
-    overdriven_network_entry = metadata.network_core.MetaNode.get_first_network_entry(control_obj, metadata.network_core.OverDrivenControl)
+    overdriven_network_entry = metadata.meta_network_utils.get_first_network_entry(control_obj, metadata.network_core.OverDrivenControl)
     addon_control = None
     if overdriven_network_entry:
         addon_network = overdriven_network_entry.get_upstream(metadata.network_core.AddonCore)
@@ -501,7 +501,7 @@ def get_component_network_list(obj_list):
     '''
     Returns a list of all component network objects from a list of scene objects
     '''
-    control_list = [x for x in obj_list if metadata.meta_properties.get_properties([x], metadata.meta_properties.ControlProperty)]
+    control_list = [x for x in obj_list if metadata.meta_property_utils.get_properties([x], metadata.meta_properties.ControlProperty)]
     component_network_list = []
     for control in control_list:
         component_network_list.extend(rigging.skeleton.get_primary_rig_network(control))
@@ -520,9 +520,9 @@ def get_control_info(control):
     Returns:
         ControlInfo. ControlInfo storing all information from the control's ControlProperty, or None
     '''
-    control_property = metadata.meta_properties.get_property(control, metadata.meta_properties.ControlProperty)
+    control_property = metadata.meta_property_utils.get_property(control, metadata.meta_properties.ControlProperty)
 
-    component_network = metadata.network_core.MetaNode.get_first_network_entry(control, metadata.network_core.RigComponent)
+    component_network = metadata.meta_network_utils.get_first_network_entry(control, metadata.network_core.RigComponent)
 
     if control_property:
         control_type = control_property.data.get('control_type')
@@ -541,12 +541,12 @@ def get_control_from_info(control_info, skeleton_dict):
 
     found_control = None
     region_root = skeleton_dict[control_info.side][control_info.region]['root']
-    component_core = metadata.network_core.MetaNode.get_first_network_entry(region_root, metadata.network_core.ComponentCore)
+    component_core = metadata.meta_network_utils.get_first_network_entry(region_root, metadata.network_core.ComponentCore)
     if component_core:
         control_list = component_core.get_downstream(metadata.network_core.ControlJoints).get_connections()
 
         for rig_control in control_list:
-            control_property = metadata.meta_properties.get_property(rig_control, metadata.meta_properties.ControlProperty)
+            control_property = metadata.meta_property_utils.get_property(rig_control, metadata.meta_properties.ControlProperty)
             if control_property.get('control_type') == control_info.control_type and control_property.get('ordered_index') == control_info.ordered_index:
                 found_control = rig_control
 
@@ -565,7 +565,7 @@ def get_control_vars_strings(rig_control):
             attr_name = attr.name().rsplit(".", 1)[-1]
             modified_dict[attr_name+".isLocked"] = True
 
-    control_property = metadata.meta_properties.get_property(rig_control, metadata.meta_properties.ControlProperty)
+    control_property = metadata.meta_property_utils.get_property(rig_control, metadata.meta_properties.ControlProperty)
     selection_lock = control_property.get("locked")
     if selection_lock:
         modified_dict["locked.selectionLock"] = True
@@ -652,7 +652,7 @@ def get_rig_control_selection():
     selection = pm.ls(sl=True)
     control_list = []
     for obj in selection:
-        control_property = metadata.meta_properties.get_property(obj, metadata.meta_properties.ControlProperty)
+        control_property = metadata.meta_property_utils.get_property(obj, metadata.meta_properties.ControlProperty)
         if control_property:
             control_list.append(obj)
