+++ Freeform.Python/V1PyCore/V1PyCore.pyproj
@@ -6,12 +6,12 @@
     <ProjectGuid>03600953-2c6a-4ff7-9656-1c17ddc464c5</ProjectGuid>
     <ProjectHome>.</ProjectHome>
     <StartupFile>__root__.py</StartupFile>
-    <SearchPath>..\..\..\Program Files\Autodesk\Maya2016\Python\Lib\site-packages\;W:\Program Files\Autodesk\Maya2020\Python\Lib\site-packages</SearchPath>
+    <SearchPath>..\..\..\Program Files\Autodesk\Maya2016\Python\Lib\site-packages\;W:\Program Files\Autodesk\Maya2020\Python\Lib\site-packages;W:\Program Files\Autodesk\Maya2022\Python37\Lib\site-packages</SearchPath>
     <WorkingDirectory>.</WorkingDirectory>
     <OutputPath>.</OutputPath>
     <Name>V1PyCore</Name>
     <RootNamespace>V1PyCore</RootNamespace>
-    <InterpreterId>Global|PythonCore|2.7</InterpreterId>
+    <InterpreterId>Global|PythonCore|3.7</InterpreterId>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
     <DebugSymbols>true</DebugSymbols>
@@ -118,6 +118,7 @@
   </ItemGroup>
   <ItemGroup>
     <InterpreterReference Include="Global|PythonCore|2.7" />
+    <InterpreterReference Include="Global|PythonCore|3.7" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="maya_standalone\" />
